scripts/requirements.txt
# For fetch_acaps_inform_severity_csv.py (run from project root)
requests>=2.28.0
python-dotenv>=1.0.0
